updated TODOs + solved warning conversions

+++ 2D Particle System/Motor2D/Particle.cpp
@@ -25,10 +25,10 @@ void Particle::Update(float dt)
 	life--;
 
 	pState.pLive.ageRatio = (float)this->life / (float)this->start_life;
-	pState.pLive.alpha = (int)(pState.pLive.ageRatio * 255.0f);
+	pState.pLive.alpha = (pState.pLive.ageRatio * 255.0f);
 	pState.pLive.radius = pState.pLive.start_radius * pState.pLive.ageRatio;
 	pState.pLive.rectSize.w = pState.pLive.rectSize.h = pState.pLive.radius;
-	pState.pLive.alpha = (uint)(pState.pLive.ageRatio * 255.0f);
+	pState.pLive.alpha = (pState.pLive.ageRatio * 255.0f);
 
 	pState.pLive.pos.x += pState.pLive.vel.x * dt;
 	pState.pLive.pos.y += pState.pLive.vel.y * dt;
@@ -36,12 +36,12 @@ void Particle::Update(float dt)
 
 void Particle::Draw()
 {
-	SDL_Rect rectTest = { pState.pLive.start_radius , pState.pLive.start_radius };
-	int centerX = pState.pLive.pos.x + ((rectTest.w - pState.pLive.rectSize.w) / 2);
-	int centerY = pState.pLive.pos.y + ((rectTest.h - pState.pLive.rectSize.h) / 2);
+	SDL_Rect rectTest = { (int)pState.pLive.start_radius, (int)pState.pLive.start_radius };
+	float centerX = pState.pLive.pos.x + ((rectTest.w - pState.pLive.rectSize.w) / 2);
+	float centerY = pState.pLive.pos.y + ((rectTest.h - pState.pLive.rectSize.h) / 2);
 
 	// SDL_Texture* texture, float alpha, int x, int y, const SDL_Rect* section, const SDL_Rect* rectSize, float speed, double angle, int pivot_x, int pivot_y
-	App->render->BlitParticle(App->psystem->GetParticleAtlas(), pState.pLive.alpha, centerX, centerY, &pState.pLive.pRect, &pState.pLive.rectSize);
+	App->render->BlitParticle(App->psystem->GetParticleAtlas(), pState.pLive.alpha, (int)centerX, (int)centerY, &pState.pLive.pRect, &pState.pLive.rectSize);
 }
 
 bool Particle::IsAlive()
